+++ src/frontend/WorldRenderer.h
@@ -22,7 +22,7 @@ class ChunksRenderer;
 class Shader;
 class Frustum;
 class Engine;
-class Chunks;
+class ChunksMatrix;
 class LevelFrontend;
 class Skybox;
 class PostProcessing;
@@ -38,7 +38,7 @@ class WorldRenderer {
     std::unique_ptr<Skybox> skybox;
     std::unique_ptr<Batch3D> batch3d;
     bool drawChunk(size_t index, Camera* camera, Shader* shader, bool culling);
-    void drawChunks(Chunks* chunks, Camera* camera, Shader* shader);
+    void drawChunks(ChunksMatrix* chunks, Camera* camera, Shader* shader);
 
     /// @brief Render level without diegetic interface
     /// @param context graphics context
