+++ src/world/Level.h
@@ -14,7 +14,6 @@ inline constexpr int DEF_PLAYER_INVENTORY_SIZE = 40;
 class Content;
 class World;
 class Player;
-class Chunks;
 class Inventory;
 class Inventories;
 class LevelEvents;
@@ -22,19 +21,19 @@ class Lighting;
 class PhysicsSolver;
 class ChunksStorage;
 
+class ChunksMatrix;
 /// @brief A level, contains chunks and objects
 class Level {
 public:
     std::unique_ptr<World> world;
     const Content* const content;
     std::vector<std::shared_ptr<Object>> objects;
-    std::unique_ptr<Chunks> chunks;
+    std::unique_ptr<ChunksMatrix> chunks;
     std::unique_ptr<ChunksStorage> chunksStorage;
     std::unique_ptr<Inventories> inventories;
 
     std::unique_ptr<PhysicsSolver> physics;
     std::unique_ptr<Lighting> lighting;
-    std::unique_ptr<LevelEvents> events;
 
     const EngineSettings& settings;
